Move CaptivePortalDetector to src/chomponents/captive_portal
BUG=364894
For trivial header file renaming in c/b and c/test:

R=mmenke@chromium.org, ygorshenin@chromium.org
TBR=sky@chromium.org

Review URL: https://codereview.chromium.org/242483003

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@266920 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: b1a160ded85df2e7911553e81900b7cc5f53f034

+++ components/captive_portal/captive_portal_testing_utils.cc
@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#include "chrome/browser/captive_portal/testing_utils.h"
+#include "components/captive_portal/captive_portal_testing_utils.h"
 
 #include "base/logging.h"
 #include "base/memory/ref_counted.h"
@@ -14,9 +14,8 @@ namespace {
 
 scoped_refptr<net::HttpResponseHeaders> CreateResponseHeaders(
     const std::string& response_headers) {
-  std::string raw_headers =
-      net::HttpUtil::AssembleRawHeaders(response_headers.c_str(),
-                                        response_headers.length());
+  std::string raw_headers = net::HttpUtil::AssembleRawHeaders(
+      response_headers.c_str(), static_cast<int>(response_headers.length()));
   return new net::HttpResponseHeaders(raw_headers);
 }
